+++ src/plugin/amulet/editor/_main.py
@@ -23,7 +23,7 @@ from amulet.app.style import set_style
 from amulet.app.invoke import invoke
 
 from . import __path__ as editor_plugin_path
-from ._window import init_and_show_editor, get_amulet_editor
+from ._main_window import init_and_show_editor, get_amulet_editor
 from ._init import init_editor_tools
 
 log = logging.getLogger(__name__)
