+++ src/plugin/amulet/editor/_main_window.py
