--- src/plugin/amulet/editor/_window/__init__.py
@@ -1,1 +1,0 @@
-from ._main_window import init_and_show_editor, get_amulet_editor
