+++ src/plugin/amulet/editor/api.py
@@ -1,4 +1,3 @@
 """This is the public interface to interact with the editor."""
 
-from ._window import get_amulet_editor
-from .widget import TabWidget, register_tab_widget
+from ._main_window import get_amulet_editor
